prepflowv1/requirements.txt
flask==3.0.0
flask-session==0.5.0
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.8
